[SERVER] create windows_integration.cpp for the server
probably deprecated as soon as flex does this
