+++ inexor/server/windows_integration.cpp
@@ -1,0 +1,266 @@
+
+#ifdef WIN32
+#ifdef STANDALONE
+
+#include "inexor/shared/cube_types.hpp"
+#include "inexor/shared/cube_unicode.hpp"
+#include "inexor/shared/cube_loops.hpp"
+#include "inexor/shared/cube_vector.hpp"
+#include "inexor/shared/cube_queue.hpp"
+
+#include "inexor/util/Logging.hpp"
+
+#include <windows.h>
+#include <shellapi.h>
+
+
+extern void fatal(const char *s, ...);
+
+
+#define LOGSTRLEN 512
+#define IDI_ICON1 1
+
+static string apptip = "";
+static HINSTANCE appinstance = NULL;
+static ATOM wndclass = 0;
+static HWND appwindow = NULL, conwindow = NULL;
+static HICON appicon = NULL;
+static HMENU appmenu = NULL;
+static HANDLE outhandle = NULL;
+static const int MAXLOGLINES = 200;
+struct logline { int len; char buf[LOGSTRLEN]; };
+static queue<logline, MAXLOGLINES> loglines;
+
+static void cleanupsystemtray()
+{
+    NOTIFYICONDATA nid;
+    memset(&nid, 0, sizeof(nid));
+    nid.cbSize = sizeof(nid);
+    nid.hWnd = appwindow;
+    nid.uID = IDI_ICON1;
+    Shell_NotifyIcon(NIM_DELETE, &nid);
+}
+
+static bool setupsystemtray(UINT uCallbackMessage)
+{
+    NOTIFYICONDATA nid;
+    memset(&nid, 0, sizeof(nid));
+    nid.cbSize = sizeof(nid);
+    nid.hWnd = appwindow;
+    nid.uID = IDI_ICON1;
+    nid.uCallbackMessage = uCallbackMessage;
+    nid.uFlags = NIF_MESSAGE | NIF_ICON | NIF_TIP;
+    nid.hIcon = appicon;
+    strcpy(nid.szTip, apptip);
+    if(Shell_NotifyIcon(NIM_ADD, &nid) != TRUE)
+        return false;
+    atexit(cleanupsystemtray);
+    return true;
+}
+
+#if 0
+static bool modifysystemtray()
+{
+    NOTIFYICONDATA nid;
+    memset(&nid, 0, sizeof(nid));
+    nid.cbSize = sizeof(nid);
+    nid.hWnd = appwindow;
+    nid.uID = IDI_ICON1;
+    nid.uFlags = NIF_TIP;
+    strcpy(nid.szTip, apptip);
+    return Shell_NotifyIcon(NIM_MODIFY, &nid) == TRUE;
+}
+#endif
+
+static void cleanupwindow()
+{
+    if(!appinstance) return;
+    if(appmenu)
+    {
+        DestroyMenu(appmenu);
+        appmenu = NULL;
+    }
+    if(wndclass)
+    {
+        UnregisterClass(MAKEINTATOM(wndclass), appinstance);
+        wndclass = 0;
+    }
+}
+
+static BOOL WINAPI consolehandler(DWORD dwCtrlType)
+{
+    switch(dwCtrlType)
+    {
+        case CTRL_C_EVENT:
+        case CTRL_BREAK_EVENT:
+        case CTRL_CLOSE_EVENT:
+            exit(EXIT_SUCCESS);
+            return TRUE;
+    }
+    return FALSE;
+}
+
+static void writeline(logline &line)
+{
+    static uchar ubuf[512];
+    size_t len = strlen(line.buf), carry = 0;
+    while(carry < len)
+    {
+        size_t numu = encodeutf8(ubuf, sizeof(ubuf), &((uchar *)line.buf)[carry], len - carry, &carry);
+        DWORD written = 0;
+        WriteConsole(outhandle, ubuf, numu, &written, NULL);
+    }
+}
+
+static void setupconsole()
+{
+    if(conwindow) return;
+    if(!AllocConsole()) return;
+    SetConsoleCtrlHandler(consolehandler, TRUE);
+    conwindow = GetConsoleWindow();
+    SetConsoleTitle(apptip);
+    //SendMessage(conwindow, WM_SETICON, ICON_SMALL, (LPARAM)appicon);
+    SendMessage(conwindow, WM_SETICON, ICON_BIG, (LPARAM)appicon);
+    outhandle = GetStdHandle(STD_OUTPUT_HANDLE);
+    CONSOLE_SCREEN_BUFFER_INFO coninfo;
+    GetConsoleScreenBufferInfo(outhandle, &coninfo);
+    coninfo.dwSize.Y = MAXLOGLINES;
+    SetConsoleScreenBufferSize(outhandle, coninfo.dwSize);
+    SetConsoleCP(CP_UTF8);
+    SetConsoleOutputCP(CP_UTF8);
+    loopv(loglines) writeline(loglines[i]);
+}
+
+enum
+{
+    MENU_OPENCONSOLE = 0,
+    MENU_SHOWCONSOLE,
+    MENU_HIDECONSOLE,
+    MENU_EXIT
+};
+
+static LRESULT CALLBACK handlemessages(HWND hWnd, UINT uMsg, WPARAM wParam, LPARAM lParam)
+{
+    switch(uMsg)
+    {
+        case WM_APP:
+            SetForegroundWindow(hWnd);
+            switch(lParam)
+            {
+                //case WM_MOUSEMOVE:
+                //	break;
+                case WM_LBUTTONUP:
+                case WM_RBUTTONUP:
+                {
+                    POINT pos;
+                    GetCursorPos(&pos);
+                    TrackPopupMenu(appmenu, TPM_CENTERALIGN|TPM_BOTTOMALIGN|TPM_RIGHTBUTTON, pos.x, pos.y, 0, hWnd, NULL);
+                    PostMessage(hWnd, WM_NULL, 0, 0);
+                    break;
+                }
+            }
+            return 0;
+        case WM_COMMAND:
+            switch(LOWORD(wParam))
+            {
+                case MENU_OPENCONSOLE:
+                    setupconsole();
+                    if(conwindow) ModifyMenu(appmenu, 0, MF_BYPOSITION|MF_STRING, MENU_HIDECONSOLE, "Hide Console");
+                    break;
+                case MENU_SHOWCONSOLE:
+                    ShowWindow(conwindow, SW_SHOWNORMAL);
+                    ModifyMenu(appmenu, 0, MF_BYPOSITION|MF_STRING, MENU_HIDECONSOLE, "Hide Console");
+                    break;
+                case MENU_HIDECONSOLE:
+                    ShowWindow(conwindow, SW_HIDE);
+                    ModifyMenu(appmenu, 0, MF_BYPOSITION|MF_STRING, MENU_SHOWCONSOLE, "Show Console");
+                    break;
+                case MENU_EXIT:
+                    PostMessage(hWnd, WM_CLOSE, 0, 0);
+                    break;
+            }
+            return 0;
+        case WM_CLOSE:
+            PostQuitMessage(0);
+            return 0;
+    }
+    return DefWindowProc(hWnd, uMsg, wParam, lParam);
+}
+
+static void setupwindow(const char *title)
+{
+    copystring(apptip, title);
+    //appinstance = GetModuleHandle(NULL);
+    if(!appinstance) fatal("failed getting application instance");
+    appicon = LoadIcon(appinstance, MAKEINTRESOURCE(IDI_ICON1));//(HICON)LoadImage(appinstance, MAKEINTRESOURCE(IDI_ICON1), IMAGE_ICON, 0, 0, LR_DEFAULTSIZE);
+    if(!appicon) spdlog::get("global")->error("failed loading icon");
+
+    appmenu = CreatePopupMenu();
+    if(!appmenu) fatal("failed creating popup menu");
+    AppendMenu(appmenu, MF_STRING, MENU_OPENCONSOLE, "Open Console");
+    AppendMenu(appmenu, MF_SEPARATOR, 0, NULL);
+    AppendMenu(appmenu, MF_STRING, MENU_EXIT, "Exit");
+    //SetMenuDefaultItem(appmenu, 0, FALSE);
+
+    WNDCLASS wc;
+    memset(&wc, 0, sizeof(wc));
+    wc.hCursor = NULL; //LoadCursor(NULL, IDC_ARROW);
+    wc.hIcon = appicon;
+    wc.lpszMenuName = NULL;
+    wc.lpszClassName = title;
+    wc.style = 0;
+    wc.hInstance = appinstance;
+    wc.lpfnWndProc = handlemessages;
+    wc.cbWndExtra = 0;
+    wc.cbClsExtra = 0;
+    wndclass = RegisterClass(&wc);
+    if(!wndclass) fatal("failed registering window class");
+
+    appwindow = CreateWindow(MAKEINTATOM(wndclass), title, 0, CW_USEDEFAULT, CW_USEDEFAULT, 0, 0, HWND_MESSAGE, NULL, appinstance, NULL);
+    if(!appwindow) fatal("failed creating window");
+
+    atexit(cleanupwindow);
+
+    if(!setupsystemtray(WM_APP)) fatal("failed adding to system tray");
+}
+
+static char *parsecommandline(const char *src, vector<char *> &args)
+{
+    char *buf = new char[strlen(src) + 1], *dst = buf;
+    for(;;)
+    {
+        while(isspace(*src)) src++;
+        if(!*src) break;
+        args.add(dst);
+        for(bool quoted = false; *src && (quoted || !isspace(*src)); src++)
+        {
+            if(*src != '"') *dst++ = *src;
+            else if(dst > buf && src[-1] == '\\') dst[-1] = '"';
+            else quoted = !quoted;
+        }
+        *dst++ = '\0';
+    }
+    args.add(NULL);
+    return buf;
+}
+
+
+int WINAPI WinMain(HINSTANCE hInst, HINSTANCE hPrev, LPSTR szCmdLine, int sw)
+{
+    vector<char *> args;
+    char *buf = parsecommandline(GetCommandLine(), args);
+    appinstance = hInst;
+
+    setupwindow("Inexor server");
+    SetPriorityClass(GetCurrentProcess(), HIGH_PRIORITY_CLASS);
+
+    int standalonemain(int argc, char **argv);
+    int status = standalonemain(args.length()-1, args.getbuf());
+
+    delete[] buf;
+    exit(status);
+    return 0;
+}
+
+#endif
+#endif
